+++ tests/unit/middleware/appStatusCheck.js
@@ -21,7 +21,7 @@ describe('appStatusCheck middleware', function() {
         this.config = new Config();
 
         this.appManager = new AppManager(this.models);
-        var app = this.app = this.appManager.buildApp(this.config);
+        var app = this.app = this.appManager.buildApp(this.config, {name: '1'});
 
         this.res = {};
         this.req = {};