+++ tests/unit/middleware/errorHandler.js
@@ -22,7 +22,7 @@ describe('errorHandler middleware', function() {
         this.config = new Config();
 
         this.appManager = new AppManager(this.models);
-        var app = this.app = this.appManager.buildApp(this.config);
+        var app = this.app = this.appManager.buildApp(this.config, {name: '1'});
 
         this.appSetStatusSpy = sinon.spy(this.app, '$setStatus');
         this.appEmitSpy = sinon.spy(this.app, 'emit');