+++ tests/unit/middleware/routeNotFound.js
@@ -20,7 +20,7 @@ describe('routeNotFound middleware', function() {
         this.config = new Config();
 
         this.appManager = new AppManager(this.models);
-        var app = this.app = this.appManager.buildApp(this.config);
+        var app = this.app = this.appManager.buildApp(this.config, {name: '1'});
 
         this.res = {};
         this.req = {};