+++ tests/unit/serviceIntegrity.js
@@ -35,7 +35,7 @@ describe('serviceIntegrity', function() {
         this.configGetStub = sinon.stub(this.config, 'get');
 
         this.appManager = new AppManager(this.models);
-        var app = this.app = this.appManager.buildApp(this.config);
+        var app = this.app = this.appManager.buildApp(this.config, {name: '1'});
     });
 
     beforeEach(function() {