+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.10.3
+
+* [FIXED] an `Error` is throwed when a `Router` is trying to register a `Route` with non-unique route `uid`. `uid` should be unique in context of an `App`
+* [FIXED] an `Error` is throwed when a App defines non-unique `name` option
+* [FIXED] each `App` presents itself independently to the `bi-depot` - this prevents service scope collision
+
 ## v0.10.2
 
 * [FIXED] `restrictByClient` looks for `bi-depot` settings under `services.privateDepot` key instead of `services.private-depot`